Make define_backend thread safe
We run Rails on a multi-threaded puma webserver. This means we need to
ensure that ratonvirus is thread safe before we use it.

We're changing the code here to avoid using a class instance variable
since this will be shared between threads.

Let's take an example where `backend_type` is set to "scanner". The
`create_#{backend_type}` method will become `create_scanner`.

This will be used to set the value of @#{backend_type} class ivar. So
```
@#{backend_type} ||= create_#{backend_type}
```
becomes
```
@scanner ||= create_scanner
```

The `create_scanner` method will return an instance of
`@#{backend_type}_defs[:klass]` which becomes `scanner_defs[:klass]`.

Let's say we're using Clamby as a virus scanner. Our @scanner class ivar
will ultimately be set like this:

```
@scanner ||= Ratonvirus::Clamby::Base.new(...)
```

The problem with assigning this in a class ivar is that this type of
object can hold different states depending on where you are in the
antivirus scanning process.

For example it holds `@errors` generated by the antivirus scanner [1].
These can get overwritten by a different thread since all threads share
the same @scanner class ivar.

So we can have a situation where:
- thread 1 sets @errors to an array of ["error1", "error2"]
- thread 2 comes along and adds an "error3" to @errors:
["error1", "error2", "error3"]
- thread 3 has just spawned and wants to set @errors to nil, as seen
here [1], which will reset it for all three threads.

To avoid sharing state between threads, we can use Thread-local
variables which can't be accessed by other threads. [2]

[1]: https://github.com/futurelearn/ratonvirus/blob/master/lib/ratonvirus/scanner/base.rb#L92
[2]: https://www.rubytapas.com/2016/11/20/ruby-thread-local-variables/

Co-authored-by: Simon Coffey <simon.coffey@futurelearn.com>