Don't allow scanners to be configured via a class
In the previous commit we changed from using a class instance variable
to using a Thread-local variable.

When we configure ratonvirus via a class, this instance variable won't
be available yet as we haven't set the Thread-local variable at this
point.

So in order to avoid using this configuration style, let's just throw a
warning letting people know it's disabled.

Co-authored-by: Simon Coffey <simon.coffey@futurelearn.com>