Throw an error if antivirus is not present
The way we add the virus scanner to a rails app is by adding an
validator to a model:

```
validates :file, antivirus: true
```

If clamav isn't available for scanning, it will [exit the validation
process early][1] without throwing an error.

This means we'll be able to upload infected files if clamav is not
running.

To put in a more stringent check, we are changing the validation process
to throw an error if clamav is not available. If a user tries to upload
a file while clamav is offline they will now be blocked from uploading.

[1]: https://github.com/futurelearn/ratonvirus/blob/9ece9882ff2f3507664411d5ad14fc39edb71e1d/app/validators/antivirus_validator.rb#L14

Co-authored-by: James Hunt <james.hunt@futurelearn.com>