Don't block uploads in development
We don't want to require developers to install clamav when they run the
project locally. So we're using the `force_availability` configuration
option which we'll set as an ENV var in our project.

This will be set to false for development and true everywhere else.

This is slightly different than what the `force_availability` option was
originally intended to do but it achieves the same effect: it's a
feature flag for requiring the antivirus to be present.

Co-authored-by: James Hunt <james.hunt@futurelearn.com>