Don't depend on `force_availability` option for setting availability
In a non-obvious way, this configuration option was previously used to
enforce the existence of an antivirus when scanning a file.

It did this by always setting `available` to true which meant in the
validation class we would continue with scanning the file.  Because the
antivirus would not be present, the process would exit with an error and
would generate a "virus found" validation error.

This is quite convoluted and we don't want the misdirection of always
setting `available` to true.

We also wanted to make the distinction between when an actual virus was
found versus not having an antivirus running. We've already handled both
of these cases in a previous commit.

Here we're removing the option to "always say we have an antivirus and
let it explode".

Co-authored-by: James Hunt <james.hunt@futurelearn.com>