Skip validation if we don't expect the antivirus to be present
We'd like to allow developers to upload files in development without
expecting them to install Clamav.

We initially assumed that `scanner.available?` will return `false` since
Clamav won't be installed locally. [1] However, calling `available?`
just checks that `Ratonvirus::Clamby::Base` is an executable class
(which is always true because it's a class present in the
ratonvirus-clamby gem). The scan will therefore still be run locally but
will fail when it doesn't find an antivirus.

So we'd actually block people from uploading when running in development
because it looks like the scanner is available.

We've set `force_availability` as a feature flag to indicate whether we
expect an antivirus to be present. Let's return early in the validation
class without scanning when we set `force_availability` to `false`.

This means we can now upload to our heart's content in development, but
still expect the antivirus to be present everywhere else.

[1]: https://github.com/futurelearn/ratonvirus/pull/3/commits/3330d91f60b522853bd2c4099b707bfdb28561ed